+++ Source/Podio .NET/Services/ItemService.cs
@@ -30,7 +30,7 @@ namespace PodioAPI.Services
         /// </param>
         /// <param name="hook">If set to false, hooks will not be executed for the change</param>
         /// <returns>Id of the created item</returns>
-        public int AddNewItem(int appId, Item item, int? spaceId = null, bool silent = false, bool hook = true)
+        public long AddNewItem(int appId, Item item, int? spaceId = null, bool silent = false, bool hook = true)
         {
             JArray fieldValues = JArray.FromObject(item.Fields.Select(f => new {external_id = f.ExternalId, field_id = f.FieldId, values = f.Values}));
 
@@ -186,7 +186,7 @@ namespace PodioAPI.Services
         ///     If true marks any new notifications on the given item as viewed, otherwise leaves any
         ///     notifications untouched, Default value true
         /// </param>
-        public Item GetItem(int itemId, bool markedAsViewed = true)
+        public Item GetItem(long itemId, bool markedAsViewed = true)
         {
             string markAsViewdValue = markedAsViewed == true ? "1" : "0";
             var requestData = new Dictionary<string, string>
@@ -208,7 +208,7 @@ namespace PodioAPI.Services
         ///     notifications untouched, Default value true
         /// </param>
         /// <returns></returns>
-        public Item GetItemBasic(int itemId, bool markedAsViewed = true)
+        public Item GetItemBasic(long itemId, bool markedAsViewed = true)
         {
             string viewedVal = markedAsViewed == true ? "1" : "0";
             var requestData = new Dictionary<string, string>
@@ -357,12 +357,12 @@ namespace PodioAPI.Services
         ///     generated
         /// </param>
         /// <returns>The id of the cloned item</returns>
-        public int CloneItem(int itemId, bool silent = false)
+        public int CloneItem(long itemId, bool silent = false)
         {
             string url = string.Format("/item/{0}/clone", itemId);
             url = Podio.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent));
             dynamic tw = _podio.Post<dynamic>(url);
-            return int.Parse(tw["item_id"].ToString());
+            return long.Parse(tw["item_id"].ToString());
         }
 
         /// <summary>
@@ -375,7 +375,7 @@ namespace PodioAPI.Services
         ///     generated
         /// </param>
         /// <param name="hook">If set to false, hooks will not be executed for the change</param>
-        public void DeleteItem(int itemId, bool silent = false, bool hook = true)
+        public void DeleteItem(long itemId, bool silent = false, bool hook = true)
         {
             string url = string.Format("/item/{0}", itemId);
             url = Podio.PrepareUrlWithOptions(url, new CreateUpdateOptions(silent, hook));
@@ -387,7 +387,7 @@ namespace PodioAPI.Services
         ///     <para>Podio API Reference: https://developers.podio.com/doc/items/delete-item-reference-7302326 </para>
         /// </summary>
         /// <param name="itemId"></param>
-        public void DeleteItemReference(int itemId)
+        public void DeleteItemReference(long itemId)
         {
             string url = string.Format("/item/{0}/ref", itemId);
             _podio.Delete<dynamic>(url);
@@ -472,7 +472,7 @@ namespace PodioAPI.Services
         /// <param name="limit">The maximum number of results to return Default value: 13</param>
         /// <param name="notItemIds">If supplied the items with these ids will not be returned</param>
         /// <returns></returns>
-        public List<Item> GetTopValuesByField(int fieldId, int limit = 13, int[] notItemIds = null)
+        public List<Item> GetTopValuesByField(int fieldId, int limit = 13, long[] notItemIds = null)
         {
             string itemIdCSV = Utilities.ArrayToCSV(notItemIds);
             var requestData = new Dictionary<string, string>
@@ -509,7 +509,7 @@ namespace PodioAPI.Services
         /// <param name="itemId"></param>
         /// <param name="fieldId"></param>
         /// <returns></returns>
-        public T GetItemFieldValues<T>(int itemId, int fieldId) where T : new()
+        public T GetItemFieldValues<T>(long itemId, int fieldId) where T : new()
         {
             string url = string.Format("/item/{0}/value/{1}", itemId, fieldId);
             return _podio.Get<T>(url);
@@ -522,7 +522,7 @@ namespace PodioAPI.Services
         /// <param name="itemId"></param>
         /// <param name="fieldId"></param>
         /// <returns></returns>
-        public Item GetItemBasicByField(int itemId, string fieldId)
+        public Item GetItemBasicByField(long itemId, string fieldId)
         {
             string url = string.Format("/item/{0}/reference/{1}/preview", itemId, fieldId);
             return _podio.Get<Item>(url);
@@ -535,7 +535,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="itemId"></param>
         /// <returns></returns>
-        public List<ItemReference> GetItemReferences(int itemId)
+        public List<ItemReference> GetItemReferences(long itemId)
         {
             string url = string.Format("/item/{0}/reference/", itemId);
             return _podio.Get<List<ItemReference>>(url);
@@ -573,7 +573,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="itemId"></param>
         /// <returns></returns>
-        public List<ItemField> GetItemValues(int itemId)
+        public List<ItemField> GetItemValues(long itemId)
         {
             string url = string.Format("/item/{0}/value", itemId);
             return _podio.Get<List<ItemField>>(url);
@@ -585,7 +585,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="itemId"></param>
         /// <returns></returns>
-        public ItemRevision GetItemRevision(int itemId, int revision)
+        public ItemRevision GetItemRevision(long itemId, int revision)
         {
             string url = string.Format("/item/{0}/revision/{1}", itemId, revision);
             return _podio.Get<ItemRevision>(url);
@@ -597,7 +597,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="itemId"></param>
         /// <returns></returns>
-        public List<ItemRevision> GetItemRevisions(int itemId)
+        public List<ItemRevision> GetItemRevisions(long itemId)
         {
             string url = string.Format("/item/{0}/revision/", itemId);
             return _podio.Get<List<ItemRevision>>(url);
@@ -626,7 +626,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="itemId"></param>
         /// <param name="status">The new status, either "invited", "accepted", "declined" or "tentative"</param>
-        public void SetParticipation(int itemId, string status)
+        public void SetParticipation(long itemId, string status)
         {
             dynamic requestData = new
             {
@@ -644,7 +644,7 @@ namespace PodioAPI.Services
         /// <param name="fieldId"></param>
         /// <param name="limit"></param>
         /// <returns></returns>
-        public List<ItemMicro> GetReferencesToItemByField(int itemId, int fieldId, int limit = 20)
+        public List<ItemMicro> GetReferencesToItemByField(long itemId, int fieldId, int limit = 20)
         {
             var requestData = new Dictionary<string, string>
             {
@@ -661,7 +661,7 @@ namespace PodioAPI.Services
         /// <param name="itemId"></param>
         /// <param name="type">The type of the reference</param>
         /// <param name="referenceId">todo: describe referenceId parameter on UpdateItemReference</param>
-        public void UpdateItemReference(int itemId, string type, int referenceId)
+        public void UpdateItemReference(long itemId, string type, int referenceId)
         {
             dynamic requestData = new
             {
@@ -680,7 +680,7 @@ namespace PodioAPI.Services
         /// <param name="itemId"></param>
         /// <param name="revisionId"></param>
         /// <returns>The id of the new revision</returns>
-        public int? RevertItemRevision(int itemId, int revisionId)
+        public int? RevertItemRevision(long itemId, int revisionId)
         {
             var url = string.Format("/item/{0}/revision/{1}", itemId, revisionId);
             var response = _podio.Delete<dynamic>(url);
@@ -697,7 +697,7 @@ namespace PodioAPI.Services
         /// <param name="itemId"></param>
         /// <param name="revisionId"></param>
         /// <returns>The id of the new revision</returns>
-        public int? RevertToRevision(int itemId, int revisionId)
+        public int? RevertToRevision(long itemId, int revisionId)
         {
             var url = string.Format("/item/{0}/revision/{1}/revert_to", itemId, revisionId);
             var response = _podio.Post<dynamic>(url);
@@ -722,7 +722,7 @@ namespace PodioAPI.Services
         /// </param>
         /// <param name="text">The text to search for. The search will be lower case, and with a wildcard in each end.</param>
         /// <returns></returns>
-        public List<Item> FindReferenceableItems(int fieldId, int limit = 13, int[] notItemIds = null,
+        public List<Item> FindReferenceableItems(int fieldId, int limit = 13, long[] notItemIds = null,
             string sort = null, string text = null)
         {
             string itemIdCSV = Utilities.ArrayToCSV(notItemIds);
@@ -743,7 +743,7 @@ namespace PodioAPI.Services
         /// </summary>
         /// <param name="itemId"></param>
         /// <returns></returns>
-        public Clone GetItemClone(int itemId)
+        public Clone GetItemClone(long itemId)
         {
             string url = string.Format("/item/{0}/clone", itemId);
             return _podio.Get<Clone>(url);
@@ -757,13 +757,13 @@ namespace PodioAPI.Services
         /// <param name="revisionFrom"></param>
         /// <param name="revisionTo"></param>
         /// <returns></returns>
-        public List<ItemDiff> GetItemRevisionDifference(int itemId, int revisionFrom, int revisionTo)
+        public List<ItemDiff> GetItemRevisionDifference(long itemId, int revisionFrom, int revisionTo)
         {
             string url = string.Format("/item/{0}/revision/{1}/{2}", itemId, revisionFrom, revisionTo);
             return _podio.Get<List<ItemDiff>>(url);
         }
 
-        public ItemCalculate Calcualate(int itemId, ItemCalculateRequest ItemCalculateRequest, int limit = 30)
+        public ItemCalculate Calcualate(long itemId, ItemCalculateRequest ItemCalculateRequest, int limit = 30)
         {
             ItemCalculateRequest.Limit = ItemCalculateRequest.Limit == 0 ? 30 : ItemCalculateRequest.Limit;
             string url = string.Format("/item/app/{0}/calculate", itemId);
